+++ src/openforms/registrations/contrib/camunda/plugin.py
@@ -18,7 +18,7 @@ from django_camunda.tasks import start_process
 from django_camunda.utils import serialize_variables
 
 from openforms.submissions.models import Submission
-from openforms.submissions.tasks import set_submission_reference
+from openforms.submissions.public_references import set_submission_reference
 
 from ...base import BasePlugin
 from ...exceptions import NoSubmissionReference, RegistrationFailed
