+++ src/openforms/registrations/contrib/demo/plugin.py
@@ -3,7 +3,7 @@ from typing import NoReturn
 from django.utils.translation import ugettext_lazy as _
 
 from openforms.submissions.models import Submission
-from openforms.submissions.tasks import set_submission_reference
+from openforms.submissions.public_references import set_submission_reference
 
 from ...base import BasePlugin
 from ...exceptions import NoSubmissionReference, RegistrationFailed
