+++ src/openforms/registrations/contrib/objects_api/plugin.py
@@ -15,7 +15,7 @@ from openforms.formio.rendering.structured import render_json
 from openforms.submissions.exports import create_submission_export
 from openforms.submissions.mapping import SKIP, FieldConf, apply_data_mapping
 from openforms.submissions.models import Submission, SubmissionReport
-from openforms.submissions.tasks import set_submission_reference
+from openforms.submissions.public_references import set_submission_reference
 from openforms.translations.utils import to_iso639_2b
 
 from ...base import BasePlugin
