+++ src/openforms/registrations/tests/test_registration_hook.py
@@ -47,7 +47,6 @@ class RegistrationHookTests(TestCase):
 
         cls.submission = SubmissionFactory.create(
             completed=True,
-            pre_registration_completed=True,
             form__registration_backend="callback",
             form__registration_backend_options={
                 "string": "some-option",
@@ -194,7 +193,7 @@ class RegistrationHookTests(TestCase):
         last_register_date = timezone.now() - timedelta(hours=1)
         submission = SubmissionFactory.create(
             last_register_date=last_register_date,
-            registration_status=RegistrationStatuses.success,
+            registration_success=True,
             form__registration_backend="callback",
             form__registration_backend_options={
                 "string": "some-option",
@@ -214,7 +213,6 @@ class RegistrationHookTests(TestCase):
     def test_submission_marked_complete_when_form_has_no_registration_backend(self):
         submission_no_registration_backend = SubmissionFactory.create(
             completed=True,
-            pre_registration_completed=True,
             form__registration_backend="",
             form__registration_backend_options={},
         )
@@ -284,7 +282,6 @@ class NumRegistrationsTest(TestCase):
     def test_limit_registration_attempts(self, mock_get_solo):
         submission = SubmissionFactory.create(
             completed=True,
-            pre_registration_completed=True,
             form__registration_backend="callback",
         )
